+++ diesel/src/pg/connection/raw.rs
@@ -43,7 +43,10 @@ impl RawConnection {
         ) };
 
         if result_ptr.is_null() {
-            Err(Error::DatabaseError(Box::new(last_error_message(self.internal_connection))))
+            Err(Error::DatabaseError(
+                DatabaseErrorKind::__Unknown,
+                Box::new(last_error_message(self.internal_connection)),
+            ))
         } else {
             unsafe {
                 Ok(PgString::new(result_ptr))