+++ diesel_tests/tests/errors.rs
@@ -1,7 +1,39 @@
 use diesel;
 use diesel::prelude::*;
 use diesel::result::Error::DatabaseError;
+use diesel::result::DatabaseErrorKind::UniqueViolation;
 use schema::*;
+
+#[test]
+fn unique_constraints_are_detected() {
+    let connection = connection();
+    diesel::insert(&User::new(1, "Sean")).into(users::table)
+        .execute(&connection).unwrap();
+
+    let failure = diesel::insert(&User::new(1, "Jim")).into(users::table)
+        .execute(&connection);
+    assert_matches!(failure, Err(DatabaseError(UniqueViolation, _)));
+}
+
+#[test]
+#[cfg(feature = "postgres")]
+fn unique_constraints_report_correct_constraint_name() {
+    let connection = connection();
+    connection.execute("CREATE UNIQUE INDEX users_name ON users (name)").unwrap();
+    diesel::insert(&User::new(1, "Sean")).into(users::table)
+        .execute(&connection).unwrap();
+
+    let failure = diesel::insert(&User::new(2, "Sean")).into(users::table)
+        .execute(&connection);
+    match failure {
+        Err(DatabaseError(UniqueViolation, e)) => {
+            assert_eq!(Some("users"), e.table_name());
+            assert_eq!(None, e.column_name());
+            assert_eq!(Some("users_name"), e.constraint_name());
+        },
+        _ => panic!("{:?} did not match Err(DatabaseError(UniqueViolation, e))", failure),
+    };
+}
 
 macro_rules! try_no_coerce {
     ($e:expr) => ({
@@ -22,7 +54,7 @@ fn cached_prepared_statements_can_be_reused_after_error() {
         try_no_coerce!(query.execute(&connection));
 
         let failure = query.execute(&connection);
-        assert_matches!(failure, Err(DatabaseError(_)));
+        assert_matches!(failure, Err(DatabaseError(UniqueViolation, _)));
         Ok(())
     });