+++ diesel_tests/tests/types_roundtrip.rs
@@ -31,7 +31,7 @@ pub fn test_type_round_trips<ST, T>(value: T) -> bool where
                 true
             }
         }
-        Err(Error::DatabaseError(ref e))
+        Err(Error::DatabaseError(_, ref e))
             if e.message() == "invalid byte sequence for encoding \"UTF8\": 0x00" => true,
         Err(e) => panic!("Query failed: {:?}", e),
     }